new changes added

+++ src/main/java/com/Task1/OodlesA1/Service/CompanyService.java
@@ -3,13 +3,13 @@ package com.Task1.OodlesA1.Service;
 import com.Task1.OodlesA1.Domain.Company;
 import com.Task1.OodlesA1.Dtos.RequestDto.CompanyDtos.CompanyCreateDto;
 import com.Task1.OodlesA1.Dtos.RequestDto.CompanyDtos.CompanyUpdateDto;
+import com.Task1.OodlesA1.Dtos.ResponseDto.GetCompanies;
 import com.Task1.OodlesA1.Exceptions.*;
 import com.Task1.OodlesA1.Repository.CompanyRepository;
-import com.Task1.OodlesA1.Repository.DepartmentRepository;
-import com.Task1.OodlesA1.Repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -30,12 +30,15 @@ public class CompanyService {
     }
 
 
-    public String deleted(Integer cId){
+    public String deleted(Integer cId) throws CompanyIsNotPresent {
         Optional<Company> company = companyRepository.findById(Long.valueOf(cId));
+        if (company.isPresent()) {
             companyRepository.deleteById(Long.valueOf(cId));
-            return "Company Deleted";
+            return "Company Deleted with Id " + cId;
+        } else {
+            throw new CompanyIsNotPresent("Company is not present with Id " + cId);
         }
-
+    }
 
     public String change(CompanyUpdateDto companyUpdateDto) throws CompanyIsNotPresent {
         Optional<Company> company = companyRepository.findById((long) Math.toIntExact(companyUpdateDto.getCid()));
@@ -51,16 +54,27 @@ public class CompanyService {
             throw new CompanyIsNotPresent("Company is not Present With given Id"+companyUpdateDto.getCid());
         }
     }
-
-    public List<Company> getList() {
-       List<Company> companyList=companyRepository.findAll();
-       return companyList;
-    }
-
+     public List<GetCompanies> getList() {
+     List<GetCompanies>getCompaniesList=new ArrayList<>();
+     List<Company>companyList=companyRepository.findAll();
+     for(Company company:companyList){
+         GetCompanies getCompanies=new GetCompanies();
+         getCompanies.setCompanyName(company.getCompanyName());
+         getCompanies.setLocation(company.getLocation());
+         getCompanies.setNoOfEmployees(company.getNoOfEmployees());
+         getCompaniesList.add(getCompanies);
+     }
+     return getCompaniesList;
+     }
 
     public Company get(Long companyId) {
         Optional<Company> company = companyRepository.findById(companyId);
             return company.get();
+    }
 
+
+    public Company getByname(String companyName) {
+      Company company=companyRepository.findByCompanyName(companyName);
+      return company;
     }
 }
